linking to criteria used to determine Yes or No
The report leads off with a bullet list of how the Chicago Council of Lawyers made their recommendations. It's a couple clicks into their site, so it wold be great to link more directly to it. Or, if linking to a PDF is no good, you could add a second page that provides the criteria list and a link to the full report.